Split native unit tests by subsystem
