+++ tests/unit/test_harness.h
@@ -1,0 +1,40 @@
+/*
+fall-of-nouraajd c++ dark fantasy game
+Copyright (C) 2025-2026  Andrzej Lis
+
+This program is free software: you can redistribute it and/or modify
+        it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+        but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+#pragma once
+
+#include <iostream>
+
+inline int failures = 0;
+
+inline void expect_true(bool condition, const char *message) {
+    if (!condition) {
+        std::cerr << "FAIL: " << message << "\n";
+        ++failures;
+    }
+}
+
+inline int finish_tests() {
+    if (failures == 0) {
+        std::cout << "All unit tests passed.\n";
+        return 0;
+    }
+
+    std::cerr << failures << " unit test(s) failed.\n";
+    return 1;
+}
